+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 - 6.6.0
   - Add JsonLogic Export for SwitchCase (PR #1013)
+  - Add support of `tooltip` for operator config (PR #1011) (issue #261)
 - 6.5.2
   - Updated dependencies. `@babel/runtime` is now dep for core package (PR #1051) (issue #964)
 - 6.5.1